huge GUI refactor
